// src/app/product/state/models/action.type.ts
export enum ProductActionTypes {
    TOGGLE_PRODUCT_CODE = 'TOGGLE_PRODUCT_CODE',
    ADD_PRODUCT = 'ADD_PRODUCT',
    SET_CURRENT_PRODUCT = 'SET_CURRENT_PRODUCT',
    UPDATE_PRODUCT = 'UPDATE_PRODUCT',
    UPDATE_PRODUCT_SUCCESS = 'UPDATE_PRODUCT_SUCCESS',
    UPDATE_PRODUCT_ERROR = 'UPDATE_PRODUCT_ERROR',
    SELECT_PRODUCT = 'SELECT_PRODUCT',
    LOAD_PRODUCTS_SUCCESS = 'LOAD_PRODUCTS_SUCCESS',
    LOAD_PRODUCTS_ERROR = 'LOAD_PRODUCTS_ERROR',
    LOAD_PRODUCTS = 'LOAD_PRODUCTS',
    ADD_PRODUCT_SUCCESS = 'ADD_PRODUCT_SUCCESS',
    ADD_PRODUCT_ERROR = 'ADD_PRODUCT_ERROR',
}